+++ NEWS.md
@@ -63,6 +63,9 @@
 * `env_get_list()` retrieves muliple bindings from an environment into
   a named list.
 
+* `with_bindings()` and `scoped_bindings()` establish temporary
+  bindings in an environment.
+
 
 # rlang 0.1.2